tild character must be escaped in hunk title ref #156
vimagit may want to search hunk position, based on hunk title.
When a hunk title has a '~' in it, the character make the search crash.
It must be escaped.

Finnally, after almost one year of open issue and several issue
contributors, I got it, thanks to akrejczinger and its docker image.

The reason I was unable to reproduce this issue is maybe that all the
vim and nvim version I used were built with another regex engine.