Fix another escaping issue, with a global escape method ref #168
Another issue with escape problem. It was finally the same culprit.

match() function take a pattern as a paramter. If a:selection[0], the
hunk title, contains any regex special character, it goes wild.

It seems safer to escape with special character \V (very nomagic)
instead of adding regularly special characters to escape() function.