fix stuck while refresh fix #170 #36
On first refresh, vimagit can be stuck during several seconds.
foldmethod=syntax seems to be the culprit.

Longer explanation:

Remove folding while we are updating the buffer. writing to the buffer
while the folding is enabled can be veryyyyy slow.
One last strange thing is that, the first time the buffer is written, it
is not slow at all. It is slow the second time, at first refresh (can be
order of seconds). Then at third time (2nd refresh), it is fast again.