Always specify diff prefix to avoid diff.noprefix errors
Fix for Issue #176 where the user has configured diff.noprefix=True and
ends up hitting an error when calling `git apply`.

Both `git diff` and `git show` are affected by the diff.noprefix
setting, and since each relies on there being a single-directory prefix
component to the filenames, make this explicit in the respective git
command calls.

To this end, always use --src-prefix=a/ and --dst-prefix=b/.  The actual
values are arbitrary and are just later stripped internally.