Use --no-prefix to generate diff
It fixes the issue #176 and the PR #177.

PR #177 was almost fine. The idea of the PR was to force src and dst
prefix. The problem was that, to generate patch for new files, we diff
the new file against /dev/null. git does not like prefix for out of
repository files.

The simple fix is to always force --no-prefix. To make it work, git
apply must use -p0 option: this option tells that there is no prefix in
the src dst files paths.