Implemented Branch Checkout (#141)
There are three bindings: <CBB>, <CBF>, and <CB->. <CBB> is the
equivalent of 'git checkout -b'; <CBF> is the equivalent of 'git
checkout -B', and <CB-> is the equivalent of 'git checkout -'.

I adapted the UI based off of the proposed UI in the issue by changing
from autocompletion(?) of typed branch names to listed branches (remote
and local alike) which can be selected with <CBB>/<CBF>, along with the
option to Close the section as well as create a new local branch.

Inline help and vimagit.txt are updated with details.