Fix jump_to function when jumping to a dirty existing buffer
jump_to opens the `filename` with `:edit`, no matter is the buffer is
already opened in vim. It fails if the buffer already exists and if it
has unsaved modifications.

Simply check if the buffer exists, and if so, open it with `:buffer`,
otherwise `:edit` it like a new file.

fixes #194