Amend loads prev. msg. even if CWD is outside repo
The path to the ".git" directory returned by
git rev-parse --git-dir
and stored in "b:magit_git_dir", can be a relative one, which would not work for opening file
"COMMIT_EDITMSG" when the currently edited file was opened from outside the
repository, so we make "b:magit_git_dir" absolute via "git rev-parse --absolute-git-dir".

Thanks to user v-slava for suggesting the use of --absolute-git-dir.